updated routes and controllers
