+++ middleware/auth.middleware.js
@@ -1,5 +1,5 @@
 const jwt = require("jsonwebtoken");
-const User = require("../models/User");
+const User = require("../models/user.Model");
 
 const requireAuth = async (req, res, next) => {
   try {
