+++ package.json
@@ -15,6 +15,8 @@
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.13.1"
+    "mongoose": "^8.13.1",
+    "nodemailer": "^6.10.0",
+    "validator": "^13.15.0"
   }
 }
