+++ server.js
@@ -2,6 +2,7 @@ require("dotenv").config();
 const express = require("express");
 const app = express();
 const bodyParser = require("body-parser");
+const interviewRoutes = require("./routes/interviewRoutes")
 const userRoutes = require("./routes/user");
 const verifyRoutes = require("./routes/verify");
 const { connect } = require("./db/connect");
@@ -21,6 +22,7 @@ app.use((req, res, next) => {
 
 app.use("/api/user", userRoutes);
 app.use("/api/verify", verifyRoutes);
+app.use("/api/interview", verifyRoutes);
 
 app.get("/", (req, res) => {
   res.send("hello from github actions");
